fix(core): report real coordinates for a negative flood-fill seed
magic_wand_with_gap_close mapped a negative seed through u32::try_from
and, on failure, built SeedOutOfBounds { x: 0, y: 0, .. } -- a valid
in-bounds coordinate that hid the actual error from anyone reading it.
The error's x/y fields are now i64 so an off-the-top-or-left seed
reports its real position. Add a test pinning the negative-seed coords.

+++ core/src/selection/error.rs
@@ -18,13 +18,15 @@ pub enum Error {
         actual_h: u32,
     },
 
-    /// Seed point for a flood-fill was outside the buffer bounds.
+    /// Seed point for a flood-fill was outside the buffer bounds. The
+    /// coordinates are signed so a negative seed (off the top or left
+    /// edge) reports its real position rather than a clamped zero.
     #[error("seed ({x}, {y}) is outside buffer bounds {width}x{height}")]
     SeedOutOfBounds {
         /// Seed x coordinate.
-        x: u32,
+        x: i64,
         /// Seed y coordinate.
-        y: u32,
+        y: i64,
         /// Buffer width.
         width: u32,
         /// Buffer height.
